Update Project 2.ipynb
Using pandas to stack the data sets of both yahoo and google stock prices and then editing the format of the date to dashes ("Timestamp": "Datetime"). Also, adding another column to the table ("stock_ID") to inform the reader which stock we are referring to (i.e. google or yahoo).